LUCENE-10313: minor clean-ups and follow-ups
